+++ src/hoc/withTheme.js
@@ -5,7 +5,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import hoistStatics from 'hoist-non-react-statics'
 import { CHANNEL } from '../models/ThemeProvider'
-import isStyledComponent from '../utils/isStyledComponent'
+import _isStyledComponent from '../utils/isStyledComponent'
 
 const wrapWithTheme = (Component: ReactClass<any>) => {
   const componentName = (
@@ -14,7 +14,7 @@ const wrapWithTheme = (Component: ReactClass<any>) => {
     'Component'
   )
 
-  const _isStyledComponent = isStyledComponent(Component)
+  const isStyledComponent = _isStyledComponent(Component)
 
   class WithTheme extends React.Component {
     static displayName = `WithTheme(${componentName})`
@@ -45,6 +45,7 @@ const wrapWithTheme = (Component: ReactClass<any>) => {
     }
 
     render() {
+      // eslint-disable-next-line react/prop-types
       const { innerRef } = this.props
       const { theme } = this.state
 
@@ -52,9 +53,10 @@ const wrapWithTheme = (Component: ReactClass<any>) => {
         <Component
           theme={theme}
           {...this.props}
-          innerRef={_isStyledComponent ? innerRef : undefined}
-          ref={_isStyledComponent ? undefined : innerRef}
+          innerRef={isStyledComponent ? innerRef : undefined}
+          ref={isStyledComponent ? undefined : innerRef}
         />
+      )
     }
   }